requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.34
botocore==1.34.34
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2023.10
cryptography==44.0.0
Django==4.2.10
django-bootstrap-datepicker-plus==5.0.5
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-storages==1.14.2
environ==1.0
gunicorn==21.2.0
idna==3.10
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dateutil==2.8.2
requests==2.32.3
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.6.0
